requirements.txt
llama-index
llama-index-embeddings-fastembed
llama-index-llms-ollama
llama-index-vector-stores-qdrant

llama-index-embeddings-ollama
llama-index-retrievers-bm25

qdrant_client
fastembed
streamlit==1.32.2
pdfplumber
tqdm
PyMuPDF
scholarly
python-dotenv